refactor: extract tmux commands, reduce ceo.py to 160 lines, clean __init__.py re-exports
- Extract tmux commands/helpers into factory/cli/_tmux_commands.py (330 lines)
- Extract flag validation, project resolution, execution into factory/cli/_ceo_helpers.py (494 lines)
- ceo.py reduced from 792 to 160 lines; cmd_ceo reduced from ~367 to ~83 lines
- Remove 21 unused private re-exports from __init__.py (32 → 11)
- Update test mock patch targets and imports to match new module locations
