+++ factory/cli/ceo.py
@@ -2,17 +2,9 @@
 from __future__ import annotations
 
 import argparse
-import hashlib
-import json
 import os
-import re
-import shlex
-import subprocess
-import structlog
 import sys
 import tempfile
-import time
-from datetime import datetime
 from pathlib import Path
 
 from factory.cli._ceo_helpers import (
@@ -28,8 +20,6 @@ from factory.cli._mode_handlers import (
     handle_review_mode,
 )
 from factory.cli._path_resolver import _resolve_focus_issue
-
-log = structlog.get_logger()
 
 
 # ── subcommand handlers ──────────────────────────────────────
@@ -188,339 +178,4 @@ def cmd_refactory(args: argparse.Namespace) -> int:
 
     os.chdir(project_path)
     os.execvp("claude", cmd)
-    return 0
-
-
-# ── tmux integration ──────────────────────────────────────────
-
-
-_TMUX_SESSION_PREFIX = "factory-"
-
-
-_TMUX_SESSIONS_FILE = Path("~/.factory/tmux_sessions.json").expanduser()
-
-
-def _tmux_session_name(project_path: Path) -> str:
-    """Derive a tmux session name from a project path."""
-    path_hash = hashlib.sha1(str(project_path).encode()).hexdigest()[:6]
-    return f"{_TMUX_SESSION_PREFIX}{project_path.name}-{path_hash}"
-
-
-def _load_tmux_session_mapping() -> dict[str, str]:
-    """Load the session→project mapping from ~/.factory/tmux_sessions.json."""
-    if _TMUX_SESSIONS_FILE.exists():
-        try:
-            return json.loads(_TMUX_SESSIONS_FILE.read_text())
-        except (json.JSONDecodeError, OSError):
-            pass
-    return {}
-
-
-def _save_tmux_session_mapping(session: str, project_path: str) -> None:
-    """Save a session→project mapping entry to ~/.factory/tmux_sessions.json."""
-    mapping = _load_tmux_session_mapping()
-    mapping[session] = project_path
-    _TMUX_SESSIONS_FILE.parent.mkdir(parents=True, exist_ok=True)
-    _TMUX_SESSIONS_FILE.write_text(json.dumps(mapping, indent=2))
-
-
-def _tmux_available() -> bool:
-    """Check if tmux is installed."""
-    try:
-        subprocess.run(["tmux", "-V"], capture_output=True, check=True)
-        return True
-    except (FileNotFoundError, subprocess.CalledProcessError):
-        return False
-
-
-def _tmux_session_alive(session: str) -> bool:
-    """Check if a tmux session exists and is alive."""
-    return (
-        subprocess.run(
-            ["tmux", "has-session", "-t", session],
-            capture_output=True,
-        ).returncode
-        == 0
-    )
-
-
-def _build_tmux_run_args(args: argparse.Namespace, project_path: Path, model: str | None) -> str:
-    """Build the 'factory ceo ...' command string from parsed args."""
-    parts = [f"factory ceo {project_path}"]
-    if args.mode:
-        parts.append(f"--mode {args.mode}")
-    if model:
-        parts.append(f"--model {shlex.quote(model)}")
-    if getattr(args, "no_github", False):
-        parts.append("--no-github")
-    if getattr(args, "profile", None):
-        parts.append(f"--profile {shlex.quote(args.profile)}")
-    if getattr(args, "focus", None):
-        parts.append(f"--focus {shlex.quote(args.focus)}")
-    if getattr(args, "refine", None):
-        parts.append(f"--refine {shlex.quote(args.refine)}")
-    if getattr(args, "clean_pr", None) is True:
-        parts.append("--clean-pr")
-    elif getattr(args, "clean_pr", None) is False:
-        parts.append("--no-clean-pr")
-    if getattr(args, "runner", None):
-        parts.append(f"--runner {shlex.quote(args.runner)}")
-    if getattr(args, "prompt", None):
-        parts.append(f"--prompt {shlex.quote(args.prompt)}")
-    if getattr(args, "branch", None):
-        parts.append(f"--branch {shlex.quote(args.branch)}")
-    if getattr(args, "min_growth", None) is not None:
-        parts.append(f"--min-growth {args.min_growth}")
-    if getattr(args, "max_new", None) is not None:
-        parts.append(f"--max-new {args.max_new}")
-    if getattr(args, "discover_only", False):
-        parts.append("--discover-only")
-    if getattr(args, "bg_agents", False):
-        parts.append("--bg-agents")
-    if getattr(args, "tmux_persist", False):
-        parts.append("--tmux-persist")
-    if getattr(args, "use_profile", False):
-        parts.append("--use-profile")
-    return " ".join(parts)
-
-
-def cmd_tmux(args: argparse.Namespace) -> int:
-    """Launch factory run inside a detached tmux session."""
-    if not _tmux_available():
-        print("Error: tmux is not installed.", file=sys.stderr)
-        return 1
-
-    project_path = Path(args.path).resolve()
-    session = args.session or _tmux_session_name(project_path)
-
-    check = subprocess.run(
-        ["tmux", "has-session", "-t", session],
-        capture_output=True,
-    )
-    if check.returncode == 0:
-        if args.attach:
-            print(f"Attaching to existing session: {session}")
-            os.execvp("tmux", ["tmux", "attach-session", "-t", session])
-        print(f"Session '{session}' already running. Use --attach or:")
-        print(f"  tmux attach -t {session}")
-        return 0
-
-    _ENV_PREFIXES = (
-        "FACTORY_",
-        "ANTHROPIC_",
-        "BOBSHELL_",
-        "OPENAI_",
-        "CODEX_",
-        "CLAUDE_CODE_",
-        "CLOUD_ML_",
-    )
-    run_cmd_parts = []
-    for key, val in sorted(os.environ.items()):
-        if key.startswith(_ENV_PREFIXES):
-            run_cmd_parts.append(f"export {key}={shlex.quote(val)}")
-    run_cmd_parts.append(f"export PATH={shlex.quote(os.environ.get('PATH', '/usr/bin'))}")
-
-    model = _resolve_model(args)
-    run_args = _build_tmux_run_args(args, project_path, model)
-    run_cmd_parts.append(run_args)
-    shell_cmd = " && ".join(run_cmd_parts)
-
-    result = subprocess.run(
-        ["tmux", "new-session", "-d", "-s", session, "-x", "200", "-y", "50", shell_cmd],
-    )
-    if result.returncode != 0:
-        print(f"Error: failed to create tmux session '{session}'", file=sys.stderr)
-        return 1
-
-    _save_tmux_session_mapping(session, str(project_path))
-
-    time.sleep(3)
-
-    if not _tmux_session_alive(session):
-        print(f"Error: session '{session}' exited immediately after launch", file=sys.stderr)
-        return 1
-
-    capture = subprocess.run(
-        ["tmux", "capture-pane", "-t", session, "-p"],
-        capture_output=True,
-        text=True,
-    )
-    if capture.returncode == 0:
-        pane_text = capture.stdout
-        _error_markers = ("Error:", "exited", "no server")
-        if any(marker in pane_text for marker in _error_markers):
-            log.warning("tmux_post_dispatch_warning", session=session)
-            print(f"Warning: session '{session}' may have errors:", file=sys.stderr)
-            for line in pane_text.strip().splitlines()[-10:]:
-                print(f"  {line}", file=sys.stderr)
-
-    print(f"Factory launched in tmux session: {session}")
-    print(f"  tmux attach -t {session}    # attach")
-    print(f"  tmux kill-session -t {session}  # stop")
-
-    if args.attach:
-        os.execvp("tmux", ["tmux", "attach-session", "-t", session])
-
-    return 0
-
-
-def cmd_tmux_ls(args: argparse.Namespace) -> int:
-    """List running factory tmux sessions."""
-    if not _tmux_available():
-        print("Error: tmux is not installed.", file=sys.stderr)
-        return 1
-
-    result = subprocess.run(
-        ["tmux", "list-sessions", "-F", "#{session_name}\t#{session_created}\t#{session_windows}"],
-        capture_output=True,
-        text=True,
-    )
-    if result.returncode != 0:
-        print("No tmux sessions running.")
-        return 0
-
-    mapping = _load_tmux_session_mapping()
-    factory_sessions = []
-    for line in result.stdout.strip().splitlines():
-        parts = line.split("\t")
-        name = parts[0]
-        if name.startswith(_TMUX_SESSION_PREFIX):
-            created = (
-                datetime.fromtimestamp(int(parts[1])).strftime("%Y-%m-%d %H:%M")
-                if len(parts) > 1
-                else "?"
-            )
-            project = mapping.get(name, "?")
-            factory_sessions.append({"session": name, "started": created, "project": project})
-
-    if not factory_sessions:
-        if getattr(args, "json_output", False):
-            print("[]")
-        else:
-            print("No factory sessions running.")
-        return 0
-
-    if getattr(args, "json_output", False):
-        print(json.dumps(factory_sessions, indent=2))
-    else:
-        print(f"{'Session':<35} {'Started':<20} {'Project'}")
-        print("-" * 80)
-        for s in factory_sessions:
-            print(f"{s['session']:<35} {s['started']:<20} {s['project']}")
-    return 0
-
-
-def cmd_tmux_capture(args: argparse.Namespace) -> int:
-    """Capture recent output from a factory tmux session."""
-    if not _tmux_available():
-        print("Error: tmux is not installed.", file=sys.stderr)
-        return 1
-
-    session = getattr(args, "session", None)
-    if not session and getattr(args, "path", None):
-        project_path = Path(args.path).resolve()
-        mapping = _load_tmux_session_mapping()
-        for s, p in mapping.items():
-            if Path(p).resolve() == project_path:
-                session = s
-                break
-        if not session:
-            session = _tmux_session_name(project_path)
-
-    if not session:
-        print("Error: specify --session or path to identify the session", file=sys.stderr)
-        return 1
-
-    if not _tmux_session_alive(session):
-        print(f"Error: session '{session}' not found", file=sys.stderr)
-        return 1
-
-    lines = getattr(args, "lines", -100)
-    result = subprocess.run(
-        ["tmux", "capture-pane", "-t", session, "-p", "-S", str(lines)],
-        capture_output=True,
-        text=True,
-    )
-    if result.returncode != 0:
-        print(f"Error: failed to capture pane for '{session}'", file=sys.stderr)
-        return 1
-
-    print(result.stdout, end="")
-    return 0
-
-
-def cmd_tmux_stop(args: argparse.Namespace) -> int:
-    """Stop a factory tmux session."""
-    if not _tmux_available():
-        print("Error: tmux is not installed.", file=sys.stderr)
-        return 1
-
-    if args.session:
-        session = args.session
-    elif args.path:
-        session = _tmux_session_name(Path(args.path).resolve())
-    elif getattr(args, "stop_all", False):
-        result = subprocess.run(
-            ["tmux", "list-sessions", "-F", "#{session_name}"],
-            capture_output=True,
-            text=True,
-        )
-        if result.returncode != 0:
-            print("No tmux sessions running.")
-            return 0
-
-        killed = 0
-        for name in result.stdout.strip().splitlines():
-            if name.startswith(_TMUX_SESSION_PREFIX):
-                subprocess.run(["tmux", "kill-session", "-t", name])
-                print(f"Stopped: {name}")
-                killed += 1
-
-        if killed == 0:
-            print("No factory sessions running.")
-        else:
-            print(f"Stopped {killed} session(s).")
-        return 0
-    else:
-        result = subprocess.run(
-            ["tmux", "list-sessions", "-F", "#{session_name}"],
-            capture_output=True,
-            text=True,
-        )
-        sessions = []
-        if result.returncode == 0:
-            for name in result.stdout.strip().splitlines():
-                if name.startswith(_TMUX_SESSION_PREFIX):
-                    sessions.append(name)
-        if sessions:
-            print("Factory sessions that would be stopped:")
-            for s in sessions:
-                print(f"  {s}")
-        else:
-            print("No factory sessions running.")
-        print("\nUse --all to stop all factory sessions.")
-        return 1
-
-    check = subprocess.run(
-        ["tmux", "has-session", "-t", session],
-        capture_output=True,
-    )
-    if check.returncode != 0:
-        print(f"Session '{session}' not found.")
-        return 1
-
-    mapping = _load_tmux_session_mapping()
-    if session not in mapping and not getattr(args, "force", False):
-        print(
-            f"Warning: session '{session}' is not in the factory session registry.",
-            file=sys.stderr,
-        )
-        print(
-            "It may not be a factory-managed session. Use --force to kill it anyway.",
-            file=sys.stderr,
-        )
-        return 1
-
-    subprocess.run(["tmux", "kill-session", "-t", session])
-    print(f"Stopped: {session}")
     return 0
